feat: make torch optional via `pip install gliner2[local]`
Move Schema, StructureBuilder, RegexValidator into a new torch-free
module (gliner2/inference/schema.py) and lazy-load torch-dependent
symbols in __init__.py via PEP 562 __getattr__.

`pip install gliner2` now only pulls pydantic, requests, urllib3, tqdm.
`pip install gliner2[local]` adds torch, transformers, safetensors,
gliner, numpy — identical to the previous default install.

All 38 existing tests pass under [local]. New regression test
(test_torch_free_import.py) verifies the bare install.

Made-with: Cursor
