--- gliner2/inference/__init__.py
@@ -1,1 +1,0 @@
-from .engine import RegexValidator, GLiNER2
