+++ pyproject.toml
@@ -21,8 +21,19 @@ maintainers = [
 ]
 
 dependencies = [
-    "gliner",
     "pydantic>=2.0.0",
+    "requests>=2.28",
+    "urllib3>=1.26",
+    "tqdm>=4.60",
 ]
 
 dynamic = ["version"]
+
+[project.optional-dependencies]
+local = [
+    "gliner",
+    "torch>=2.0",
+    "transformers>=4.40",
+    "safetensors>=0.4",
+    "numpy>=1.23",
+]
